Rename : 컴포넌트 구조할당 및 디렉토리 변경
