+++ src/Basic/Typography/index.js
@@ -1,0 +1,1 @@
+export { default as Typography } from "./Typography";
