+++ src/Components/SideBarItem/index.js
@@ -1,0 +1,1 @@
+export { default as SideBarItem } from "./SideBarItem";
