+++ src/Components/SubNavItem/index.js
@@ -1,0 +1,1 @@
+export { default as SubNavbarItem } from "./SubNavbarItem";
