+++ src/Layouts/Sidebar/index.js
@@ -1,0 +1,1 @@
+export { default as SideBar } from "./SideBar";
