+++ src/Layouts/SubNavbar/index.js
@@ -1,0 +1,1 @@
+export { default as SubNavBar } from "./SubNavBar";
