--- src/Sidebar/SidebarItem.stories.js
@@ -1,93 +1,0 @@
-import React from "react";
-import { SidebarItem } from "./SidebarItem";
-import { Sidebar } from "./SideBar";
-
-export default {
-  title: "Design System/SidebarItem",
-  component: SidebarItem,
-};
-
-export const Default = (args) => <SidebarItem {...args} />;
-
-Default.args = {
-  contents: [
-    {
-      title: "직무",
-      description: "프론트엔드",
-      icon: "briefcase",
-    },
-    {
-      title: "고용형태",
-      description: "정규직",
-      icon: "briefcase",
-    },
-    {
-      title: "경력",
-      description: "신입",
-      icon: "briefcase",
-    },
-    {
-      title: "근무위치",
-      description: "서울 강남구",
-      icon: "briefcase",
-    },
-  ],
-  badges: ["JavaScript", "TypeScript", "Vue.js", "React", "Redux", "Svelte"],
-};
-
-Default.args = {
-  contents: {
-    title: "직무",
-    description: "프론트엔드",
-    icon: "briefcase",
-  },
-};
-
-export const All = (args) => (
-  <>
-    <Sidebar theme="dark" />
-    <h1>Light Sub Navbar Item</h1>
-    <hr />
-    <p />
-    <SidebarItem {...args} />
-    <SidebarItem {...args} />
-    <SidebarItem {...args} />
-    <SidebarItem {...args} />
-    <h1> Loading Light Sub Navbar Item</h1>
-    <hr />
-    <p />
-    <SidebarItem {...args} isLoading />
-    <SidebarItem {...args} isLoading />
-    <SidebarItem {...args} isLoading />
-    <SidebarItem {...args} isLoading />
-    <h1>Dark Sub Navbar Item</h1>
-    <hr />
-    <p />
-    <SidebarItem theme="dark" {...args} />
-    <SidebarItem theme="dark" {...args} />
-    <SidebarItem theme="dark" {...args} />
-    <SidebarItem theme="dark" {...args} />
-    <h1> Loading Dark Sub Navbar Item</h1>
-    <hr />
-    <p />
-    <SidebarItem {...args} theme="dark" isLoading />
-    <SidebarItem {...args} theme="dark" isLoading />
-    <SidebarItem {...args} theme="dark" isLoading />
-    <SidebarItem {...args} theme="dark" isLoading />
-  </>
-);
-
-All.args = {
-  contents: {
-    title: "직무",
-    description: "프론트엔드",
-    icon: "briefcase",
-  },
-};
-
-export const Layout = (args) => (
-  <>
-    <Sidebar />
-    <Sidebar theme="dark" />
-  </>
-);
